Add script for contributed abstract summary score.
This script was used to compute the contributed abstract normalized summary
score.  After normalizing the reviewers scores by the mean and standard
deviation of their per-category ratings, the categories are weighted by their
importance and combined.  The average normalized score was use as the primary
ranking method in the talk/poster/rejection selection process.