http:// -> https:// links
some of these are invalid, but of the ones that still work, it's better
to just stick with https. Note that there are still some places (png,
pdf, csv, and makefiles) where there are http:// links, but those don't
end up in rendered webiste portions / manual.

This commit was performed via the following command (on a Debian
system):

    git grep -l http\: | grep -E "\.(rst|txt|md|html|in)" | xargs sed -ie "s/http:/https:/g"